AI Lead Scraper & Personalize/backend/requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.6.0

# Database & Excel Export
openpyxl==3.1.2
pandas==2.1.4
